// src/types/api.d.ts
export interface getChat_GetChat_chat {
  __typename: "Chat";
  passengerId: number;
  driverId: number;
  messages: (getChat_GetChat_chat_messages | null)[];
}

export interface getChat_GetChat {
  __typename: "GetChatResponse";
  ok: boolean;
  error: string | null;
  chat: getChat_GetChat_chat | null;
}

export interface getChat {
  GetChat: getChat_GetChat;
}

export interface getChatVariables {
  chatId: number;
}

/* tslint:disable */
/* eslint-disable */
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL mutation operation: sendMessage
// ====================================================

export interface sendMessage_SendChatMessage_message {
  __typename: "Message";
  id: number;
  text: string;
  userId: number | null;
}

export interface sendMessage_SendChatMessage {
  __typename: "SendChatMessageResponse";
  ok: boolean;
  error: string | null;
  message: sendMessage_SendChatMessage_message | null;
}

export interface sendMessage {
  SendChatMessage: sendMessage_SendChatMessage;
}

export interface sendMessageVariables {
  text: string;
  chatId: number;
}
// ...
export interface userProfile_GetMyProfile_user {
  __typename: "User";
  id: number;
  fullName: string | null;
  firstName: string;
  lastName: string;
  email: string | null;
  profilePhoto: string | null;
  isDriving: boolean;
}

export interface userProfile_GetMyProfile {
  __typename: "GetMyProfileResponse";
  ok: boolean;
  error: string | null;
  user: userProfile_GetMyProfile_user | null;
}

export interface userProfile {
  GetMyProfile: userProfile_GetMyProfile;
}